[mod] fix import path
