+++ ls_improved/lsi_visual.py
@@ -1,4 +1,4 @@
-from config import Config
+from .config import Config
 
 class LsiVisualTransforms():
     def __init__(self):
